+++ zarrs/src/array/codec/array_to_bytes/sharding/sharding_partial_decoder_async.rs
@@ -18,8 +18,8 @@ use crate::{
             AsyncByteIntervalPartialDecoder, AsyncBytesPartialDecoderTraits, CodecChain,
             CodecError, CodecOptions,
         },
-        ravel_indices, ArrayBytes, ArrayBytesFixedDisjointView, ArrayBytesRaw, ArrayIndices,
-        ChunkRepresentation, ChunkShape, DataType, DataTypeSize, RawBytesOffsets,
+        ravel_indices, ArrayBytes, ArrayBytesFixedDisjointView, ArrayBytesOffsets, ArrayBytesRaw,
+        ArrayIndices, ChunkRepresentation, ChunkShape, DataType, DataTypeSize,
     },
     array_subset::IncompatibleDimensionalityError,
     indexer::{IncompatibleIndexerError, Indexer},
@@ -635,5 +635,8 @@ async fn partial_decode_variable_indexer(
         offsets.push(bytes.len());
     }
 
-    Ok(ArrayBytes::new_vlen(bytes, RawBytesOffsets::new(offsets)?)?)
+    Ok(ArrayBytes::new_vlen(
+        bytes,
+        ArrayBytesOffsets::new(offsets)?,
+    )?)
 }