+++ zarrs/src/array/codec/codec_partial_default.rs
@@ -8,8 +8,8 @@ use zarrs_storage::{
 
 use crate::{
     array::{
-        array_bytes::update_array_bytes, ArrayBytes, ArrayBytesRaw, BytesRepresentation,
-        ChunkRepresentation, RawBytesOffsets,
+        array_bytes::update_array_bytes, ArrayBytes, ArrayBytesOffsets, ArrayBytesRaw,
+        BytesRepresentation, ChunkRepresentation,
     },
     array_subset::ArraySubset,
 };
@@ -110,7 +110,7 @@ where
             Ok(match self.decoded_representation.data_type().size() {
                 DataTypeSize::Fixed(_) => ArrayBytes::new_flen(vec![]),
                 DataTypeSize::Variable => {
-                    ArrayBytes::new_vlen(vec![], RawBytesOffsets::new(vec![0]).unwrap()).unwrap()
+                    ArrayBytes::new_vlen(vec![], ArrayBytesOffsets::new(vec![0]).unwrap()).unwrap()
                 }
             })
         }
@@ -472,7 +472,7 @@ where
             Ok(match self.decoded_representation.data_type().size() {
                 DataTypeSize::Fixed(_) => ArrayBytes::new_flen(vec![]),
                 DataTypeSize::Variable => {
-                    ArrayBytes::new_vlen(vec![], RawBytesOffsets::new(vec![0]).unwrap()).unwrap()
+                    ArrayBytes::new_vlen(vec![], ArrayBytesOffsets::new(vec![0]).unwrap()).unwrap()
                 }
             })
         }