+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking**: `ArrayBytes::validate()` now takes a `DataType` instead of a `DataTypeSize`
 - **Breaking**: Mark `ArrayBytes` as non-exhaustive
 - **Breaking**: Rename `RawBytes` to `ArrayBytesRaw` and add deprecated alias
+- **Breaking**: Rename `RawBytesOffsets` to `ArrayBytesOffsets` and add deprecated alias
 
 ### Removed